Rename avgpool and maxpool to avg_pool3d and max_pool
