+++ library/src/tensor_operation_instance/gpu/avg_pool3d_bwd/device_avg_pool3d_bwd_ndhwc_bf16_instance.cpp
@@ -1,17 +1,17 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2018-2023, Advanced Micro Devices, Inc. All rights reserved.
 
-#include "maxpool_bwd_instance_common.hpp"
+#include "avg_pool3d_bwd_ndhwc_instance_common.hpp"
 
 namespace ck {
 namespace tensor_operation {
 namespace device {
 namespace instance {
 
-void add_device_maxpool_bwd_f16_instances(
-    std::vector<std::unique_ptr<DeviceMaxPoolBwd<F16, I32, F16>>>& instances)
+void add_device_avgpool_bwd_ndhwc_f16_instances(
+    std::vector<std::unique_ptr<DeviceAvgPoolBwd<3, F16, F16, NDHWC, NDHWC>>>& instances)
 {
-    add_device_operation_instances(instances, device_maxpool_bwd_instances<F16, I32, F16>{});
+    add_device_operation_instances(instances, device_avgpool_bwd_ndhwc_f16_instances{});
 }
 
 } // namespace instance
