--- library/src/tensor_operation_instance/gpu/avgpool_bwd/CMakeLists.txt
@@ -1,11 +1,0 @@
-set(DEVICE_AVGPOOL_BWD_INSTANCES)
-if(DTYPES MATCHES "fp16" OR NOT DEFINED DTYPES)
-    list(APPEND DEVICE_AVGPOOL_BWD_INSTANCES device_avg_pool_bwd_ndhwc_f16_instance.cpp)
-endif()
-if(DTYPES MATCHES "bf16" OR NOT DEFINED DTYPES)
-    list(APPEND DEVICE_AVGPOOL_BWD_INSTANCES device_avg_pool_bwd_ndhwc_bf16_instance.cpp)
-endif()
-if(DTYPES MATCHES "fp32" OR NOT DEFINED DTYPES)
-    list(APPEND DEVICE_AVGPOOL_BWD_INSTANCES device_avg_pool_bwd_ndhwc_f32_instance.cpp)
-endif()
-add_instance_library(device_avgpool_bwd_instance ${DEVICE_AVGPOOL_BWD_INSTANCES})
