+++ library/src/tensor_operation_instance/gpu/max_pool_bwd/CMakeLists.txt
@@ -8,4 +8,4 @@ endif()
 if(DTYPES MATCHES "fp32" OR NOT DEFINED DTYPES)
     list(APPEND DEVICE_MAXPOOL_BWD_INSTANCES device_max_pool_bwd_f32_instance.cpp)
 endif()
-add_instance_library(device_maxpool_bwd_instance ${DEVICE_MAXPOOL_BWD_INSTANCES})
+add_instance_library(device_max_pool_bwd_instance ${DEVICE_MAXPOOL_BWD_INSTANCES})
