+++ library/src/tensor_operation_instance/gpu/max_pool_bwd/device_max_pool_bwd_bf16_instance.cpp
@@ -1,17 +1,17 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2018-2023, Advanced Micro Devices, Inc. All rights reserved.
 
-#include "avgpool_bwd_ndhwc_instance_common.hpp"
+#include "max_pool_bwd_instance_common.hpp"
 
 namespace ck {
 namespace tensor_operation {
 namespace device {
 namespace instance {
 
-void add_device_avgpool_bwd_ndhwc_f16_instances(
-    std::vector<std::unique_ptr<DeviceAvgPoolBwd<3, F16, F16, NDHWC, NDHWC>>>& instances)
+void add_device_maxpool_bwd_bf16_instances(
+    std::vector<std::unique_ptr<DeviceMaxPoolBwd<BF16, I32, BF16>>>& instances)
 {
-    add_device_operation_instances(instances, device_avgpool_bwd_ndhwc_f16_instances{});
+    add_device_operation_instances(instances, device_maxpool_bwd_instances<BF16, I32, BF16>{});
 }
 
 } // namespace instance
